Newest Version of 'Application' project space for SoundViz
This has the formatting for files to work straight away even though they are not defined in the same folder, and it does it in more of traditional way of importing things, so we should be able to independently build on this without anymore of the problems having to do with us agreeing how screens fit together. Also, the methodology for this one is very visible in the couple of modules that ARE already in here. You will notice, that the menu is not popping up yet in this one, but I believe it is in the old one if you cannot remember what that meu looked like. 

When you make a screen, it should recieve a pygame.surface as parameter 1, and then AppHomeFolder == '' should be the second parameter so that when you go to call your new screen in Driver.py, the code that appends the folder name onto home folder will do its things and then the new code integrated for you and everyone else that downloads.